% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAPI.R
\name{createAPI}
\alias{createAPI}
\title{Create an API for trained ensemble model}
\usage{
createAPI(host, port)
}
\arguments{
\item{host}{your device host address}

\item{port}{address of you post to host app on}
}
\value{
starts an API app to run your predictive model
}
\description{
Create an API for trained ensemble model
}
\examples{
# data("iris")
# mm = multipleModels(train = iris, test = iris, y = "Species", models = c("C5.0", "parRF"))
# ensembleModel = ensembleTrain(mm,
#                               train = iris,
#                               test = iris,
#                               y = "Species",
#                               emsembleModelTrain = "C5.0")
# predictEnsemble(ensembleModel, iris)
# saveRDS(ensembleModel, "C:/Documents/savedEnsembleModel.RDS")
# readRDS("C:/Documents/savedEnsembleModel.RDS")
}
