% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleTrain.R
\name{ensembleTrain}
\alias{ensembleTrain}
\title{Training ensemble model}
\usage{
ensembleTrain(mm, train, test, y, emsembleModelTrain, metric, nfolds,
  repeats)
}
\arguments{
\item{mm}{A trained object from mm function}

\item{train}{A training data frame}

\item{test}{A testing data frame}

\item{y}{Response variable}

\item{emsembleModelTrain}{An ensemble model to train from caret model list}

\item{metric}{A minimization metric for training. If not mentioned, for regression RMSE and for classification Kappa value will be used.}

\item{nfolds}{Number of kfolds for cross validation. By default, 5 will be used.}

\item{repeats}{Number of repeats for cross validation. By default, 5 will be used.}
}
\value{
confusion matrix of ensemble model and returns ensemble model
}
\description{
Training ensemble model
}
\examples{
# data("iris")
# ensembleTrain(mm, train = iris, test = iris, y = "Species"
#                , emsembleModelTrain = "C5.0")
}
