% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCreationF.R
\name{featureCreationF}
\alias{featureCreationF}
\title{Transforms to freqency using FFT and extracts statistical features}
\usage{
featureCreationF(data)
}
\arguments{
\item{data}{A vector of length greater than 2}
}
\value{
Returns a list of 22 statistical features for data
}
\description{
Transforms to freqency using FFT and extracts statistical features
}
\examples{
data = rnorm(50)
featureCreationF(data)
}
