% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleModels.R
\name{multipleModels}
\alias{multipleModels}
\title{Training multiple models for benchmark comparison}
\usage{
multipleModels(train, test, y, metric, nfolds, repeats, models)
}
\arguments{
\item{train}{A training data frame}

\item{test}{A testing data frame}

\item{y}{Response variable}

\item{metric}{A minimization metric for training. If not mentioned, for regression RMSE and for classification Kappa value will be used.}

\item{nfolds}{Number of kfolds for cross validation. By default, 5 will be used.}

\item{repeats}{Number of repeats for cross validation. By default, 5 will be used.}

\item{models}{A character list of models to train based on caret package structure}
}
\value{
summary of all trained models and trained models
}
\description{
Training multiple models for benchmark comparison
}
\examples{
#data("iris")
# multipleModels(train = iris, test = iris, y = "Species", models = c("C5.0", "parRF"))

## results
##      Accuracy Kappa AccuracyLower AccuracyUpper AccuracyNull AccuracyPValue McnemarPValue
##C5.0      1.00  1.00     0.9757074     1.0000000    0.3333333   2.702787e-72           NaN
##parRF     0.96  0.94     0.9149722     0.9851815    0.3333333   2.525127e-60           NaN
}
