% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numSummary.R
\name{numSummary}
\alias{numSummary}
\title{Provides descriptive statistics for a data frame}
\usage{
numSummary(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Returns a table with descriptive statistics for a data frame.
}
\description{
Provides descriptive statistics for a data frame
}
\examples{
## data(iris)
## numSummary(iris)
}
