% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictEnsemble.R
\name{predictEnsemble}
\alias{predictEnsemble}
\title{predict for ensemble}
\usage{
predictEnsemble(ensembleModel, test)
}
\arguments{
\item{ensembleModel}{trained ensemble model}

\item{test}{a test dataframe to do predictions}
}
\value{
a dataframe with response variables
}
\description{
predict for ensemble
}
\examples{
# data("iris")
# mm = multipleModels(train = iris, test = iris, y = "Species", models = c("C5.0", "parRF"))
# ensembleModel = ensembleTrain(mm,
#                               train = iris,
#                               test = iris,
#                               y = "Species",
#                               emsembleModelTrain = "C5.0")
# predictEnsemble(ensembleModel, iris)
}
